+++ src/GalaxyWiki.API/DTOs/CelestialBodyUpdateRequest.cs
@@ -1,0 +1,9 @@
+namespace GalaxyWiki.API.DTOs
+{
+    public class CelestialBodyUpdateRequest
+    {
+        public string BodyName { get; set; } = string.Empty;
+        public int? BodyTypeId { get; set; }
+        public int? OrbitsId { get; set; }
+    }
+}
